more space-efficient hypergraph
This version of the hypergraph data structure reserves some additional
space for the incidence structure of representative hypernode.
That way, we can avoid unnecessary growing/resizing of the incidence
array. Choosing the node with the most free space however seems to
affect partitioning quality. Since it is currently not clear why that
happens, we do not use this version of the data structure as default.